Extract the the getFeatured function into the utils file
